requirements.txt
piexif==1.1.3
pillow==10.0.0